move secrets factory to its own file
