+++ lib/openstax_aws.rb
@@ -85,6 +85,7 @@ require "openstax/aws/distribution"
 require "openstax/aws/change_set"
 require "openstax/aws/secrets_specification"
 require "openstax/aws/secrets"
+require "openstax/aws/secrets_factory"
 require "openstax/aws/stack"
 require "openstax/aws/stack_factory"
 require "openstax/aws/deployment_base"
